requirements.txt
# ============================
# Build essentials
# ============================
setuptools>=68.0.0
wheel>=0.41.0
grpcio>=1.59.0

# ============================
# Core backend
# ============================
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
pydantic-core==2.14.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.27.2
aiohttp==3.9.1
requests==2.31.0
anyio==3.7.1
asgiref==3.11.0

# ============================
# Numerical core (LOCKED)
# ============================
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2

# ============================
# ML / Transformers
# ============================
torch==2.1.2
torchvision==0.16.2
transformers==4.34.0
tokenizers==0.14.1
sentence-transformers==2.2.2
tiktoken==0.5.2
huggingface-hub==0.17.3
accelerate==0.30.1
safetensors==0.4.3

# ============================
# RAG stack
# ============================
chromadb==0.5.3
langchain==0.1.0
langchain-core==0.1.23
langchain-community==0.0.10
langsmith==0.0.87
pypdf==3.17.4
beautifulsoup4==4.12.2
orjson==3.10.7
mmh3==4.1.0

# ============================
# Vector DB
# ============================
qdrant-client==1.7.3

# ============================
# CV / Media (NO NumPy v2)
# ============================
opencv-python==4.9.0.80
pillow==10.4.0
matplotlib==3.9.2

# ============================
# Audio / Speech
# ============================
faster-whisper==1.0.3
ctranslate2==4.6.2
soundfile==0.13.1
pydub==0.25.1
librosa==0.11.0
audioread==3.1.0
soxr==1.0.0
sounddevice==0.5.3
phonemizer-fork==3.3.2
espeakng-loader==0.2.4
kokoro==0.9.4

# ============================
# WebRTC / Streaming
# ============================
# aiortc==1.14.0
av>=11.0.0
aioice==0.10.2
pylibsrtp==1.0.0
pyee==13.0.0
ifaddr==0.2.0

# ============================
# NLP / spaCy
# ============================
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
thinc==8.3.10
srsly==2.4.8
cymem==2.0.8
murmurhash==1.0.10
preshed==3.0.9
blis==1.3.3
catalogue==2.0.10
confection==0.1.5
wasabi==1.1.3
weasel==0.4.3
cloudpathlib==0.20.0
smart_open==7.0.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# ============================
# Databases & Auth
# ============================
motor==3.6.0
pymongo==4.9.2
SQLAlchemy==2.0.35
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0

# ============================
# Google / OAuth
# ============================
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httplib2==0.22.0
cachetools==5.5.0
rsa==4.9
pyasn1==0.6.1
pyasn1-modules==0.4.1

# ============================
# Utilities
# ============================
groq==0.9.0
loguru==0.7.2
rich==13.7.1
tenacity==8.5.0
joblib==1.4.2
regex==2024.9.11
tqdm==4.66.5
filelock==3.15.4
packaging==23.2
platformdirs==4.2.2
psutil==6.0.0
PyYAML==6.0.3

# ============================
# PDF / Reporting
# ============================
reportlab==4.0.9

# ============================
# Networking / HTTP internals
# ============================
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
h11==0.16.0
httptools==0.7.1
watchfiles==1.1.1
websockets==12.0
websocket-client==1.9.0
multidict==6.1.0
yarl==1.11.1
frozenlist==1.4.1
aiosignal==1.3.1
attrs==24.2.0
# mediapipe removed: requires protobuf<4 which conflicts with grpcio-tools>=1.59
